Reuse anyenv init output
I don't want to execute anyenv init command as much as possible because
the execution time is about 0.4 seconds when is powered from battery or
about 0.14 seconds when ac adaptor is plugged. The time is a cost for
fast login shell starting.

No need for anyenv init executing per every login shells if the output
is saved because the output is always same if new env is not installed.
So, save anyenv init executing output and load it in login shell
starting. The execution time is about 0.005 seconds!

But the effect is so few because login shell is executed only once.